+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/Query/Actor/GetSelectionActorsQueryCommand.cs
@@ -13,14 +13,8 @@
 // limitations under the License.
 
 using System;
+using MediatR;
 
-namespace Energinet.DataHub.MarketParticipant.Application.Commands.User
-{
-    public sealed record UserOverviewItemDto(
-        Guid Id,
-        string Email,
-        string Name,
-        string? PhoneNumber,
-        DateTimeOffset CreatedDate,
-        bool Active);
-}
+namespace Energinet.DataHub.MarketParticipant.Application.Commands.Query.Actor;
+
+public sealed record GetSelectionActorsQueryCommand(Guid UserId) : IRequest<GetSelectionActorsQueryResponse>;