+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/Query/Actor/SelectionActorDto.cs
@@ -12,21 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketParticipant.Client
-{
-    /// <summary>
-    /// BFF client for Energinet.DataHub.MarketParticipant.
-    /// </summary>
-    public interface IMarketParticipantClient :
-        IMarketParticipantOrganizationClient,
-        IMarketParticipantUserClient,
-        IMarketParticipantActorClient,
-        IMarketParticipantGridAreaClient,
-        IMarketParticipantActorContactClient,
-        IMarketParticipantGridAreaOverviewClient,
-        ITokenClient,
-        IMarketParticipantUserOverviewClient,
-        IMarketParticipantActorQueryClient
-    {
-    }
-}
+using System;
+
+namespace Energinet.DataHub.MarketParticipant.Application.Commands.Query.Actor;
+
+public sealed record SelectionActorDto(Guid Id, string Gln, string Name);