+++ source/Energinet.DataHub.MarketParticipant.Common/RepositoryRegistration.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.MarketParticipant.Domain.Repositories;
-using Energinet.DataHub.MarketParticipant.Domain.Repositories.Slim;
+using Energinet.DataHub.MarketParticipant.Domain.Repositories.Query;
 using Energinet.DataHub.MarketParticipant.Infrastructure.Persistence.Repositories;
 using SimpleInjector;
 
@@ -32,10 +32,10 @@ namespace Energinet.DataHub.MarketParticipant.Common
             container.Register<IGridAreaOverviewRepository, GridAreaOverviewRepository>(Lifestyle.Scoped);
             container.Register<IUserOverviewRepository, UserOverviewRepository>(Lifestyle.Scoped);
             container.Register<IGridAreaAuditLogEntryRepository, GridAreaAuditLogEntryRepository>(Lifestyle.Scoped);
-            container.Register<Domain.Repositories.IUserRepository, UserRepository>(Lifestyle.Scoped);
+            container.Register<IUserRepository, UserRepository>(Lifestyle.Scoped);
             container.Register<IUserRoleTemplateRepository, UserRoleTemplateRepository>(Lifestyle.Scoped);
-            container.Register<IActorRepository, Infrastructure.Persistence.Repositories.Slim.ActorRepository>(Lifestyle.Scoped);
-            container.Register<Domain.Repositories.Slim.IUserRepository, Infrastructure.Persistence.Repositories.Slim.UserRepository>(Lifestyle.Scoped);
+            container.Register<IActorQueryRepository, Infrastructure.Persistence.Repositories.Query.ActorQueryRepository>(Lifestyle.Scoped);
+            container.Register<IUserQueryRepository, Infrastructure.Persistence.Repositories.Query.UserQueryRepository>(Lifestyle.Scoped);
             container.Register<IUserIdentityRepository, UserIdentityRepository>(Lifestyle.Scoped);
             container.Register<IUserRoleAssignmentAuditLogEntryRepository, UserRoleAssignmentAuditLogEntryRepository>(Lifestyle.Scoped);
         }